Update: split structure_selection in import layers
