+++ enzy_htp/structure/structure_selection/__init__.py
@@ -3,5 +3,4 @@
 Author: Qianzhen (QZ) Shao, <shaoqz@icloud.com>
 Date: 2023-02-15
 """
-from .general import *
-from .stru_selection import StruSelection
+from .general import select_stru
